fix "successful Addition" Alert
the alert was launched when addition and edting operation done
